refactor(stt): move TranscriptionResult to base
